working profile, fix bugs

+++ src/ws/server.ts
@@ -46,6 +46,8 @@ const findClient = (id: number) => {
 const handleInvite = async (from: User, toId: number) => {
 	const to = await prisma.user.findUnique({ where: { id: toId } });
 	if (!to) return;
+  const sentInvites = await prisma.invite.findMany({where: {fromId: from.id, toId}})
+  if (sentInvites.length) return;
 	const invite = await prisma.invite.create({ data: { fromId: from.id, toId: to.id } });
 	const recipient = findClient(toId);
 	if (recipient)
@@ -66,12 +68,16 @@ const handleReject = async (from: User, inviteId: number) => {
 const handleConfirm = async (socket: Socket, from: User, inviteId: number) => {
 	const invite = await prisma.invite.findUnique({ where: { id: inviteId } });
 	if (!invite || invite.toId !== from.id) return;
+	const isBusy =
+		(await prisma.user.findUnique({ where: { id: invite.fromId } }))?.gameId !== null ||
+		(await prisma.user.findUnique({ where: { id: invite.toId } }))?.gameId !== null;
+	if (isBusy) return;
 	await prisma.invite.delete({ where: { id: inviteId } });
 	const recipient = findClient(invite.fromId);
 	const game = await prisma.game.create({
 		data: {
-      whitePlayerId: invite.fromId,
-      blackPlayerId: invite.toId,
+			whitePlayerId: invite.fromId,
+			blackPlayerId: invite.toId,
 			FEN: 'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1'
 		}
 	});
@@ -79,6 +85,22 @@ const handleConfirm = async (socket: Socket, from: User, inviteId: number) => {
 	await prisma.user.update({ where: { id: invite.toId }, data: { gameId: game.id } });
 	socket.emit('confirmation', { from });
 	if (recipient) recipient.socket.emit('confirmation', { from });
+};
+
+const handleSurrender = async (from: User) => {
+	const game = await prisma.game.findUnique({ where: { id: from.gameId || -1 } });
+	if (!game) return;
+  const blackClient = findClient(game.blackPlayerId);
+  const whiteClient = findClient(game.whitePlayerId);
+	const winnerId = game.blackPlayerId !== from.id ? game.blackPlayerId : game.whitePlayerId;
+	await prisma.game.update({
+		where: { id: game.id },
+		data: { loserId: from.id, winnerId, endedAt: new Date() }
+	});
+	await prisma.user.update({ where: { id: game.blackPlayerId }, data: { gameId: null } });
+	await prisma.user.update({ where: { id: game.whitePlayerId }, data: { gameId: null } });
+  blackClient?.socket?.emit('surrender');
+  whiteClient?.socket?.emit('surrender');
 };
 
 export const webSocketServer = {
@@ -93,24 +115,24 @@ export const webSocketServer = {
 			socket.on('invite', async (toId) => await handleInvite(user, toId));
 			socket.on('reject', async (inviteId) => await handleReject(user, inviteId));
 			socket.on('confirm', async (inviteId) => await handleConfirm(socket, user, inviteId));
+      socket.on('surrender', async () => await handleSurrender(user));
 
-      // implementing game itself
-      socket.on('move', async (move) => {
-        const game = await prisma.game.findUnique({where: {id: user.gameId || -1}});
-        if (!game) return;
-        const chess = new Chess(game.FEN);
-        try {
-          chess.move(move);
-          const fen = chess.fen();
-          await prisma.game.update({where: {id: game.id}, data: {FEN: fen}})
-          const blackPlayer = findClient(game.blackPlayerId);
-          const whitePlayer = findClient(game.whitePlayerId);
-          blackPlayer?.socket?.emit('move', fen);
-          whitePlayer?.socket?.emit('move', fen);
-        }
-        catch {};
-      });
-      // implementing game itself
+			// implementing game itself
+			socket.on('move', async (move) => {
+				const game = await prisma.game.findUnique({ where: { id: user.gameId || -1 } });
+				if (!game) return;
+				const chess = new Chess(game.FEN);
+				try {
+					chess.move(move);
+					const fen = chess.fen();
+					await prisma.game.update({ where: { id: game.id }, data: { FEN: fen } });
+					const blackPlayer = findClient(game.blackPlayerId);
+					const whitePlayer = findClient(game.whitePlayerId);
+					blackPlayer?.socket?.emit('move', fen);
+					whitePlayer?.socket?.emit('move', fen);
+				} catch {}
+			});
+			// implementing game itself
 		});
 	}
 };
